fix(virtio): 修复 PCI VirtIO 队列通知和调试改进
关键修复：
1. 修复队列通知地址计算 - VirtIOPCI::get_notify_addr()
   - 对于 PCI VirtIO，队列通知地址 = notify_cfg_bar + (queue_index * notify_off_multiplier)
   - 队列 0 的通知地址 = 0x40003000（而不是 0x40003050）

2. 修复 VirtQueue::notify() 使用 u16 写入
   - VirtIO 1.0 规范：队列通知是 16 位写入
   - 之前错误地使用 u32，现在使用 u16

3. 添加 PCI VirtIO 设备的调试输出
   - 在 virtio_pci.rs::read_block() 中添加通知地址打印
   - 在 init.rs 中添加 PCI VirtIO I/O 尝试

4. 添加 PCI VirtIO 设备的全局存储和访问
   - VIRTIO_PCI_BLK 静态变量
   - register_pci_device() 和 get_pci_device() 函数
   - init.rs 优先尝试从 PCI VirtIO 设备读取

调试改进：
- 添加队列通知地址的调试打印
- 添加响应状态读取的打印
- 添加 PCI VirtIO I/O 尝试的详细日志

影响的文件：
- kernel/src/drivers/virtio/virtio_pci.rs: 修复通知地址，添加调试
- kernel/src/drivers/virtio/queue.rs: 修复 notify() 使用 u16
- kernel/src/drivers/virtio/mod.rs: 添加 PCI 设备存储
- kernel/src/init.rs: 添加 PCI VirtIO I/O 尝试
- kernel/src/drivers/virtio/probe.rs: 注册 PCI 设备

+++ kernel/src/drivers/virtio/queue.rs
@@ -192,9 +192,10 @@ impl VirtQueue {
         // 内存屏障：确保所有队列更新对设备可见
         core::sync::atomic::fence(core::sync::atomic::Ordering::Release);
         unsafe {
-            let queue_notify = self.queue_notify as *mut u32;
+            let queue_notify = self.queue_notify as *mut u16;
             // 写入队列索引（对于第一个队列是 0）
-            core::ptr::write_volatile(queue_notify, 0);
+            // VirtIO 1.0 规范：队列通知是 16 位写入
+            core::ptr::write_volatile(queue_notify, 0u16);
         }
     }
